# specs/aps-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "APS Content and Automation API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/content/briefing/daily": {
      "get": {
        "summary": "Get Daily Briefing",
        "operationId": "get_daily_briefing_v1_content_briefing_daily_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyBriefingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource or materialized content not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Artifact expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Invalid content or internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/content/projects": {
      "get": {
        "summary": "Get Project Catalog",
        "operationId": "get_project_catalog_v1_content_projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCatalogResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource or materialized content not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Artifact expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Invalid content or internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/content/projects/{project_id}/briefing": {
      "get": {
        "summary": "Get Project Briefing",
        "operationId": "get_project_briefing_v1_content_projects__project_id__briefing_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9-]{0,63}$",
              "title": "Project Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectBriefingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource or materialized content not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Artifact expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Invalid content or internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/content/services/maintenance": {
      "get": {
        "summary": "Get Service Maintenance",
        "operationId": "get_service_maintenance_v1_content_services_maintenance_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "due",
                "overdue",
                "upcoming",
                "all"
              ],
              "type": "string",
              "default": "due",
              "title": "Scope"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceMaintenanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource or materialized content not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Artifact expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Invalid content or internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/content/status": {
      "get": {
        "summary": "Get Content Status",
        "operationId": "get_content_status_v1_content_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Operation forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource or materialized content not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Artifact expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Invalid content or internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Dependency not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentStatusItem": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "missing",
              "refreshing",
              "ready",
              "failed"
            ],
            "title": "Status"
          },
          "formats": {
            "items": {
              "type": "string",
              "enum": [
                "json",
                "html"
              ]
            },
            "type": "array",
            "title": "Formats"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "succeeded": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Succeeded",
            "default": 0
          },
          "failed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failed",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ContentStatusItem"
      },
      "ContentStatusResponse": {
        "properties": {
          "vault_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vault Commit"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "generator_version": {
            "type": "string",
            "title": "Generator Version"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          },
          "content": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ContentStatusItem"
            },
            "type": "object",
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "generator_version",
          "stale",
          "content"
        ],
        "title": "ContentStatusResponse"
      },
      "DailyBriefingData": {
        "properties": {
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "summary": {
            "$ref": "#/components/schemas/DailySummary"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectBriefing"
            },
            "type": "array",
            "title": "Projects"
          },
          "service_maintenance": {
            "items": {
              "$ref": "#/components/schemas/ServiceMaintenanceItem"
            },
            "type": "array",
            "title": "Service Maintenance"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of_date",
          "summary"
        ],
        "title": "DailyBriefingData"
      },
      "DailyBriefingResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "content_type": {
            "type": "string",
            "const": "daily_briefing",
            "title": "Content Type"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "vault_commit": {
            "type": "string",
            "pattern": "^[0-9a-f]{7,64}$",
            "title": "Vault Commit"
          },
          "generator_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Generator Version"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "partial_failure": {
            "type": "boolean",
            "title": "Partial Failure",
            "default": false
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "data": {
            "$ref": "#/components/schemas/DailyBriefingData"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content_type",
          "generated_at",
          "vault_commit",
          "generator_version",
          "sha256",
          "data"
        ],
        "title": "DailyBriefingResponse"
      },
      "DailySummary": {
        "properties": {
          "active_projects": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Active Projects"
          },
          "tasks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tasks"
          },
          "decisions": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Decisions"
          },
          "services_due": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Services Due"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "active_projects",
          "tasks",
          "decisions",
          "services_due"
        ],
        "title": "DailySummary"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "ProjectBriefingData": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectBriefing"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectBriefingData"
      },
      "ProjectBriefingResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "content_type": {
            "type": "string",
            "const": "project_briefing",
            "title": "Content Type"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "vault_commit": {
            "type": "string",
            "pattern": "^[0-9a-f]{7,64}$",
            "title": "Vault Commit"
          },
          "generator_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Generator Version"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "partial_failure": {
            "type": "boolean",
            "title": "Partial Failure",
            "default": false
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "data": {
            "$ref": "#/components/schemas/ProjectBriefingData"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content_type",
          "generated_at",
          "vault_commit",
          "generator_version",
          "sha256",
          "data"
        ],
        "title": "ProjectBriefingResponse"
      },
      "ProjectCatalogData": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectCatalogItem"
            },
            "type": "array",
            "title": "Projects"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProjectCatalogData"
      },
...
      "ProjectCatalogResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "content_type": {
            "type": "string",
            "const": "project_catalog",
            "title": "Content Type"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "vault_commit": {
            "type": "string",
            "pattern": "^[0-9a-f]{7,64}$",
            "title": "Vault Commit"
          },
          "generator_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Generator Version"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "partial_failure": {
            "type": "boolean",
            "title": "Partial Failure",
            "default": false
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "data": {
            "$ref": "#/components/schemas/ProjectCatalogData"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content_type",
          "generated_at",
          "vault_commit",
          "generator_version",
          "sha256",
          "data"
        ],
        "title": "ProjectCatalogResponse"
      },
...
      "ServiceMaintenanceData": {
        "properties": {
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "scope": {
            "type": "string",
            "enum": [
              "due",
              "overdue",
              "upcoming",
              "all"
            ],
            "title": "Scope"
          },
          "summary": {
            "$ref": "#/components/schemas/MaintenanceSummary"
          },
          "services": {
            "items": {
              "$ref": "#/components/schemas/ServiceMaintenanceItem"
            },
            "type": "array",
            "title": "Services"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of_date",
          "scope",
          "summary"
        ],
        "title": "ServiceMaintenanceData"
      },
...
      "ServiceMaintenanceResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "content_type": {
            "type": "string",
            "const": "service_maintenance",
            "title": "Content Type"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "vault_commit": {
            "type": "string",
            "pattern": "^[0-9a-f]{7,64}$",
            "title": "Vault Commit"
          },
          "generator_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Generator Version"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "partial_failure": {
            "type": "boolean",
            "title": "Partial Failure",
            "default": false
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          },
          "data": {
            "$ref": "#/components/schemas/ServiceMaintenanceData"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content_type",
          "generated_at",
          "vault_commit",
          "generator_version",
          "sha256",
          "data"
        ],
        "title": "ServiceMaintenanceResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}